style: Extract the `Position` structure to a file
